Refactor: Extract edge detection logic to utility function
